+++ packages/core/src/statement.js
@@ -87,7 +87,6 @@ function exists(ezs, pluginName) {
     return pluginsList[pluginName] !== undefined;
 }
 
-
 export default {
     get,
     set,