+++ packages/lodex/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.5](https://github.com/Inist-CNRS/ezs/compare/@ezs/lodex@1.5.4...@ezs/lodex@1.5.5) (2020-07-28)
+
+
+### Bug Fixes
+
+* 🐛 missing file ([34000d8](https://github.com/Inist-CNRS/ezs/commit/34000d8825f169f229f4dd2ea7eba7d331cf4b2b))
+
+
+
+
+
+## [1.5.4](https://github.com/Inist-CNRS/ezs/compare/@ezs/lodex@1.5.3...@ezs/lodex@1.5.4) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/lodex
+
+
+
+
+
+## [1.5.3](https://github.com/Inist-CNRS/ezs/compare/@ezs/lodex@1.5.2...@ezs/lodex@1.5.3) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/lodex
+
+
+
+
+
 ## [1.5.2](https://github.com/Inist-CNRS/ezs/compare/@ezs/lodex@1.5.1...@ezs/lodex@1.5.2) (2020-06-12)
 
 **Note:** Version bump only for package @ezs/lodex