+++ packages/loterre/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.8](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.7...@ezs/loterre@1.0.8) (2020-07-30)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.7](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.6...@ezs/loterre@1.0.7) (2020-07-29)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.6](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.5...@ezs/loterre@1.0.6) (2020-07-28)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.5](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.4...@ezs/loterre@1.0.5) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.4](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.3...@ezs/loterre@1.0.4) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.3](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.2...@ezs/loterre@1.0.3) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
+## [1.0.2](https://github.com/Inist-CNRS/ezs/compare/@ezs/loterre@1.0.1...@ezs/loterre@1.0.2) (2020-07-27)
+
+**Note:** Version bump only for package @ezs/loterre
+
+
+
+
+
 ## 1.0.1 (2020-07-02)
 
 **Note:** Version bump only for package @ezs/loterre